# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Opportunity Verifier API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface AnalyzeRequest {
  url: string;
}

export type AnalyzeTextRequestInputType =
  (typeof AnalyzeTextRequestInputType)[keyof typeof AnalyzeTextRequestInputType];

export const AnalyzeTextRequestInputType = {
  email: "email",
  text: "text",
  image: "image",
} as const;

export interface AnalyzeTextRequest {
  text: string;
  inputType?: AnalyzeTextRequestInputType;
}

export interface AnalyzeImageRequest {
  /** Base64-encoded image (data URI or raw base64) */
  imageBase64: string;
}

export type ScamFlagSeverity =
  (typeof ScamFlagSeverity)[keyof typeof ScamFlagSeverity];

export const ScamFlagSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface ScamFlag {
  category: string;
  severity: ScamFlagSeverity;
  message: string;
}

export interface AnalyzeResult {
  url: string;
  trustScore: number;
  grade: string;
  flags: ScamFlag[];
  sslValid: boolean;
  domainAgeDays?: number | null;
  domainExtension: string;
  inputFieldCount: number;
  scamKeywordsFound: string[];
  summary: string;
}

export interface TextAnalysisResult {
  trustScore: number;
  grade: string;
  flags: ScamFlag[];
  scamKeywordsFound: string[];
  summary: string;
  inputType: string;
  extractedText?: string | null;
}

export interface ErrorResponse {
  error: string;
  detail?: string;
}