# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Student Opportunity Verifier API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: analyze
    description: URL / Text / Image analysis
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /analyze:
    post:
      operationId: analyzeUrl
      tags: [analyze]
      summary: Analyze a URL for scam indicators
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeRequest"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyzeResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /analyze/text:
    post:
      operationId: analyzeText
      tags: [analyze]
      summary: Analyze pasted email or social media text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeTextRequest"
      responses:
        "200":
          description: Text analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TextAnalysisResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /analyze/image:
    post:
      operationId: analyzeImage
      tags: [analyze]
      summary: Analyze an image (OCR + scam scan)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalyzeImageRequest"
      responses:
        "200":
          description: Image OCR analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TextAnalysisResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    AnalyzeRequest:
      type: object
      properties:
        url:
          type: string
      required: [url]

    AnalyzeTextRequest:
      type: object
      properties:
        text:
          type: string
        inputType:
          type: string
          enum: [email, text, image]
          default: text
      required: [text]

    AnalyzeImageRequest:
      type: object
      properties:
        imageBase64:
          type: string
          description: Base64-encoded image (data URI or raw base64)
      required: [imageBase64]

    ScamFlag:
      type: object
      properties:
        category:
          type: string
        severity:
          type: string
          enum: [low, medium, high]
        message:
          type: string
      required: [category, severity, message]

    AnalyzeResult:
      type: object
      properties:
        url:
          type: string
        trustScore:
          type: number
        grade:
          type: string
        flags:
          type: array
          items:
            $ref: "#/components/schemas/ScamFlag"
        sslValid:
          type: boolean
        domainAgeDays:
          type: number
          nullable: true
        domainExtension:
          type: string
        inputFieldCount:
          type: number
        scamKeywordsFound:
          type: array
          items:
            type: string
        summary:
          type: string
      required:
        - url
        - trustScore
        - grade
        - flags
        - sslValid
        - domainExtension
        - inputFieldCount
        - scamKeywordsFound
        - summary

    TextAnalysisResult:
      type: object
      properties:
        trustScore:
          type: number
        grade:
          type: string
        flags:
          type: array
          items:
            $ref: "#/components/schemas/ScamFlag"
        scamKeywordsFound:
          type: array
          items:
            type: string
        summary:
          type: string
        inputType:
          type: string
        extractedText:
          type: string
          nullable: true
      required:
        - trustScore
        - grade
        - flags
        - scamKeywordsFound
        - summary
        - inputType

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        detail:
          type: string
      required: [error]