# lib/api-zod/src/generated/types/analyzeImageRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Opportunity Verifier API
 * OpenAPI spec version: 0.2.0
 */

export interface AnalyzeImageRequest {
  /** Base64-encoded image (data URI or raw base64) */
  imageBase64: string;
}