# lib/api-zod/src/generated/types/analyzeTextRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Opportunity Verifier API
 * OpenAPI spec version: 0.2.0
 */
import type { AnalyzeTextRequestInputType } from "./analyzeTextRequestInputType";

export interface AnalyzeTextRequest {
  text: string;
  inputType?: AnalyzeTextRequestInputType;
}