# lib/api-zod/src/generated/types/scamFlag.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Opportunity Verifier API
 * OpenAPI spec version: 0.2.0
 */
import type { ScamFlagSeverity } from "./scamFlagSeverity";

export interface ScamFlag {
  category: string;
  severity: ScamFlagSeverity;
  message: string;
}