# lib/api-zod/src/generated/types/textAnalysisResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Student Opportunity Verifier API
 * OpenAPI spec version: 0.2.0
 */
import type { ScamFlag } from "./scamFlag";

export interface TextAnalysisResult {
  trustScore: number;
  grade: string;
  flags: ScamFlag[];
  scamKeywordsFound: string[];
  summary: string;
  inputType: string;
  extractedText?: string | null;
}